Cloud Computing for Scientific Applications
This project presents a solution for the execution of biological
software that allows to reveal the complex embryonic development that
use fluorescence correlation spectroscopy under the cloud computer
paradigm.

Microsoft Azure technologies were used to implement a highly scalable
and accessible architecture.

The implementation makes use of high performance algorithms such as a
load balancer or parametric sweep on distributed systems.

For more information about the project please visit the project's site.